@@ -14,57 +14,38 @@
   },
   "readmeFilename": "README.md",
   "engines": {
-    "node": ">=6.5"
+    "node": ">=7.5"
   },
   "dependencies": {
+    "@types/fs-extra": "0.0.37",
+    "@types/node": "^7.0.5",
     "expand-home-dir": "0.0.3",
     "fs-extra": "^2.0.0",
     "lodash": "^4.17.4",
-    "object-assign": "^4.1.1",
-    "util.pkgwrap": "0.0.5",
-    "util.rstrip": "0.0.7",
+    "util.rstrip": "0.0.10",
     "yargs": "^6.6.0"
   },
   "devDependencies": {
-    "ava": "^0.18.1",
-    "coveralls": "^2.11.15",
+    "@types/mocha": "^2.2.39",
+    "coveralls": "^2.11.16",
+    "mocha": "^3.2.0",
     "nyc": "^10.1.2",
-    "xo": "^0.17.1"
+	  "tslint": "^4.5.1",
+	  "typescript": "^2.2.1",
+    "util.pkgwrap": "0.0.11"
   },
-  "main": "index.js",
+  "main": "./index.js",
+  "types": "./index.d.ts",
   "bin": {
     "leader": "./cli.js"
   },
   "scripts": {
-    "postinstall": "./node_modules/.bin/pkgwrap --postinstall",
-    "test": "./node_modules/.bin/pkgwrap --testing",
-    "report": "./node_modules/.bin/pkgwrap --reporting",
-    "coverage": "./node_modules/.bin/pkgwrap --coverage"
-  },
-  "xo": {
-    "env": {
-      "browser": true,
-      "node": true,
-      "es6": true
-    },
-    "global": [
-      "window",
-      "describe",
-      "it",
-      "JSON"
-    ],
-    "rules": {
-      "space-before-function-paren": "off",
-      "unicorn/filename-case": [
-        "error",
-        {
-          "case": "snakeCase"
-        }
-      ],
-      "arrow-parens": [
-        "error",
-        "as-needed"
-      ]
-    }
+	  "all": "npm run lint && npm install && npm run build && npm test && npm run report",
+	  "build": "./node_modules/.bin/pkgwrap --build",
+	  "coverage": "./node_modules/.bin/pkgwrap --coverage",
+	  "lint": "./node_modules/.bin/pkgwrap --lint",
+	  "postinstall": "./node_modules/.bin/pkgwrap --postinstall",
+	  "report": "./node_modules/.bin/pkgwrap --reporting",
+	  "test": "./node_modules/.bin/pkgwrap --testing"
   }
 }
